src/requirements.txt
torch
torchvision
efficientnet_pytorch
faiss-cpu
pillow
